fix bug

+++ src/aircraft.cpp
@@ -129,7 +129,7 @@ bool Aircraft::move()
         }
         else
         {
-            if (--fuel == 0)
+            if (--fuel == 0.f)
             {
                 std::cout << "Aircraft : " << flight_number << " will crash !" << std::endl;
                 return false;
@@ -178,14 +178,15 @@ bool Aircraft::is_circling() const
     return false;
 }
 
-bool Aircraft::is_low_on_fuel() const {
+bool Aircraft::is_low_on_fuel() const
+{
     return fuel < 200;
 }
 
-void Aircraft::refill(int& fuel_stock)
+void Aircraft::refill(float& fuel_stock)
 {
-    int fuel_needed = MAX_FUEL - fuel;
-    int fuel_to_fill = fuel_stock > fuel_needed ? fuel_stock : fuel_needed ;
+    int fuel_needed  = MAX_FUEL - fuel;
+    int fuel_to_fill = fuel_stock > fuel_needed ? fuel_stock : fuel_needed;
     fuel += fuel_to_fill;
     fuel_stock -= fuel_to_fill;
     std::cout << "Flight number " << flight_number << " filled with " << fuel_to_fill << std::endl;
